restructuring mnle/categorical estimators v0.0
