+++ sbi/neural_nets/density_estimators/__init__.py
@@ -1,3 +1,10 @@
 from sbi.neural_nets.density_estimators.base import DensityEstimator
+from sbi.neural_nets.density_estimators.categorical_net import (
+    CategoricalMassEstimator,
+    CategoricalNet,
+)
+from sbi.neural_nets.density_estimators.mixed_density_estimator import (
+    MixedDensityEstimator,
+)
 from sbi.neural_nets.density_estimators.nflows_flow import NFlowsFlow
 from sbi.neural_nets.density_estimators.zuko_flow import ZukoFlow
